+++ CHANGELOG.md
@@ -4,8 +4,8 @@ Changelog
 ### Version 1.1
 Released on ???
 
-- Now loads all the photos in Romain's photostream
-- Basic behaviour, no customization, shows last 100 photos in stream
+- NEW: loads all the photos in Romain's photostream
+- NEW: updates the photos list only once a day, to save on network usage
 
 ### Version 1.0
 Released on 20th February 2014